lab2/pyproject.toml
[project]
name = "lab2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "pycryptodome>=3.23.0",
]
